requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
ca-certificates=2023.05.30=hca03da5_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
fonttools=4.42.0=pypi_0
importlib-resources=6.0.1=pypi_0
kiwisolver=1.4.4=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
matplotlib=3.7.2=pypi_0
ncurses=6.4=h313beb8_0
numpy=1.24.4=pypi_0
openssl=3.0.10=h1a28f6b_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py38hca03da5_0
praat-parselmouth=0.4.3=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.17=hb885b13_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
readline=8.2=h1a28f6b_0
seaborn=0.12.2=pypi_0
setuptools=68.0.0=py38hca03da5_0
six=1.16.0=pypi_0
sqlite=3.41.2=h80987f9_0
tk=8.6.12=hb8d0fd4_0
tzdata=2023.3=pypi_0
wheel=0.38.4=py38hca03da5_0
xz=5.4.2=h80987f9_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5a0b063_0
